refactor: déplacer BGGSearch + bggApi vers features/bgg/
